jarvis-ai-assistant/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
speechrecognition==3.10.0
pyttsx3==2.90
pyautogui==0.9.54
pyaudio==0.2.14
loguru==0.7.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0

# Windows-specific (for volume control)
pypiwin32==223; platform_system=="Windows"